show warning for missing BlendShapeClip for Perfect Sync
