+++ VMagicMirrorConfig/VMagicMirrorConfig/Model/InterprocessCommunication/ReceiveMessageNames.cs
@@ -16,6 +16,7 @@
         public const string MidiNoteOn = nameof(MidiNoteOn);
 
         public const string ExTrackerCalibrateComplete = nameof(ExTrackerCalibrateComplete);
+        public const string ExTrackerSetPerfectSyncMissedClipNames = nameof(ExTrackerSetPerfectSyncMissedClipNames);
 
         public const string VRoidModelLoadCompleted = nameof(VRoidModelLoadCompleted);
         public const string VRoidModelLoadCanceled = nameof(VRoidModelLoadCanceled);
