+++ VMagicMirrorConfig/VMagicMirrorConfig/Resources/Japanese.xaml
@@ -68,9 +68,11 @@
     <sys:String x:Key="ExTracker_Enable">外部トラッキングを有効化</sys:String>
     <sys:String x:Key="ExTracker_Enable_ImageLipSync">リップシンクも外部トラッキングの値を適用</sys:String>    
 
-    <sys:String x:Key="ExTracker_Enable_PerfectSync">パーフェクトシンクで表情を操作</sys:String>
+    <sys:String x:Key="ExTracker_Enable_PerfectSync">パーフェクトシンクを使用</sys:String>
     <sys:String x:Key="ExTracker_PerfectSync_UseVRoidDefault">VRoid用のデフォルト設定を使う(※VRoidモデルでのみ動作)</sys:String>
     <sys:String x:Key="ExTracker_HowTo_PerfectSync">パーフェクトシンクとは? (Webページを開きます)</sys:String>
+    <sys:String x:Key="ExTracker_PerfectSync_MissingBlendShape_Header">モデルのセットアップを確認して下さい！ </sys:String>
+    <sys:String x:Key="ExTracker_PerfectSync_MissingBlendShape_HeaderSub">(クリックで詳細)</sys:String>
 
     <sys:String x:Key="ExTracker_Calibrate">現在位置で顔をキャリブレーション</sys:String>
